+++ mbuild/lib/bulk_materials/amorphous_silica_bulk.pdb
